requirements.txt
comtypes
fastapi
tqdm
requests
uvicorn
sqlalchemy
passlib
bcrypt
python-multipart
python-jose